# PythonApi/openapi.json
{"openapi": "3.1.0", "info": {"title": "Product Inventory API", "description": "API for managing product inventory with categories", "version": "v1"}, "paths": {"/": {"get": {"summary": "Redirect To Swagger", "operationId": "redirect_to_swagger__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}, "/api/ProductCategories": {"get": {"tags": ["Product Categories"], "summary": "Get Product Categories", "operationId": "GetProductCategories", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"items": {"$ref": "#/components/schemas/ProductCategory"}, "type": "array", "title": "Response Getproductcategories"}}}}}}, "post": {"tags": ["Product Categories"], "summary": "Create Product Category", "operationId": "CreateProductCategory", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateProductCategoryCommand"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "integer", "title": "Response Createproductcategory"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/ProductCategories/{id}": {"get": {"tags": ["Product Categories"], "summary": "Get Product Category", "operationId": "GetProductCategory", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ProductCategory"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Product Categories"], "summary": "Update Product Category", "operationId": "UpdateProductCategory", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateProductCategoryCommand"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Product Categories"], "summary": "Delete Product Category", "operationId": "DeleteProductCategory", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/Products": {"get": {"tags": ["Products"], "summary": "Get Products", "operationId": "GetProducts", "parameters": [{"name": "category_id", "in": "query", "required": false, "schema": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Category Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Product"}, "title": "Response Getproducts"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "post": {"tags": ["Products"], "summary": "Create Product", "operationId": "CreateProduct", "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CreateProductCommand"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"type": "integer", "title": "Response Createproduct"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/Products/{id}": {"get": {"tags": ["Products"], "summary": "Get Product", "operationId": "GetProduct", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Product"}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["Products"], "summary": "Update Product", "operationId": "UpdateProduct", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UpdateProductCommand"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["Products"], "summary": "Delete Product", "operationId": "DeleteProduct", "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"CreateProductCategoryCommand": {"properties": {"name": {"type": "string", "title": "Name"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}}, "type": "object", "required": ["name"], "title": "CreateProductCategoryCommand"}, "CreateProductCommand": {"properties": {"sku": {"type": "string", "title": "Sku"}, "name": {"type": "string", "title": "Name"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "category_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Category Id"}, "stock": {"type": "integer", "title": "Stock", "default": 0}, "price": {"anyOf": [{"type": "number"}, {"type": "string"}], "title": "Price"}}, "type": "object", "required": ["sku", "name", "price"], "title": "CreateProductCommand"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "Product": {"properties": {"id": {"type": "integer", "title": "Id"}, "sku": {"type": "string", "title": "Sku"}, "name": {"type": "string", "title": "Name"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "category_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Category Id"}, "category": {"anyOf": [{"$ref": "#/components/schemas/ProductCategory"}, {"type": "null"}]}, "stock": {"type": "integer", "title": "Stock", "default": 0}, "price": {"type": "string", "title": "Price"}}, "type": "object", "required": ["id", "sku", "name", "price"], "title": "Product"}, "ProductCategory": {"properties": {"id": {"type": "integer", "title": "Id"}, "name": {"type": "string", "title": "Name"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}}, "type": "object", "required": ["id", "name"], "title": "ProductCategory"}, "UpdateProductCategoryCommand": {"properties": {"name": {"type": "string", "title": "Name"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}}, "type": "object", "required": ["name"], "title": "UpdateProductCategoryCommand"}, "UpdateProductCommand": {"properties": {"sku": {"type": "string", "title": "Sku"}, "name": {"type": "string", "title": "Name"}, "description": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Description"}, "category_id": {"anyOf": [{"type": "integer"}, {"type": "null"}], "title": "Category Id"}, "stock": {"type": "integer", "title": "Stock"}, "price": {"anyOf": [{"type": "number"}, {"type": "string"}], "title": "Price"}}, "type": "object", "required": ["sku", "name", "stock", "price"], "title": "UpdateProductCommand"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}}}